# Optimise matching process using duckdb