--- a/src/copairs/map/average_precision.py
+++ b/src/copairs/map/average_precision.py
@@ -8,7 +8,7 @@ import numpy as np
 import pandas as pd
 
 from copairs import compute
-from copairs.matching import Matcher, UnpairedException
+from copairs.matching import UnpairedException, find_pairs
 
 from .filter import evaluate_and_filter, flatten_str_list, validate_pipeline_input
 
@@ -178,37 +178,19 @@ def average_precision(
     # Reset metadata index for consistent indexing
     meta = meta.reset_index(drop=True).copy()
 
-    # Initialize the Matcher object to find pairs based on metadata rules
     logger.info("Indexing metadata...")
-    matcher = Matcher(meta, columns, seed=0)
 
     # Identify positive pairs based on `pos_sameby` and `pos_diffby`
     logger.info("Finding positive pairs...")
-    pos_pairs = matcher.get_all_pairs(sameby=pos_sameby, diffby=pos_diffby)
-    pos_total = sum(len(p) for p in pos_pairs.values())
-    if pos_total == 0:
+    pos_pairs = find_pairs(meta, sameby=pos_sameby, diffby=pos_diffby)
+    if len(pos_pairs) == 0:
         raise UnpairedException("Unable to find positive pairs.")
-
-    # Convert positive pairs to a NumPy array for efficient computation
-    pos_pairs = np.fromiter(
-        itertools.chain.from_iterable(pos_pairs.values()),
-        dtype=np.dtype((np.uint32, 2)),
-        count=pos_total,
-    )
 
     # Identify negative pairs based on `neg_sameby` and `neg_diffby`
     logger.info("Finding negative pairs...")
-    neg_pairs = matcher.get_all_pairs(sameby=neg_sameby, diffby=neg_diffby)
-    neg_total = sum(len(p) for p in neg_pairs.values())
-    if neg_total == 0:
+    neg_pairs = find_pairs(meta, sameby=neg_sameby, diffby=neg_diffby)
+    if len(neg_pairs) == 0:
         raise UnpairedException("Unable to find negative pairs.")
-
-    # Convert negative pairs to a NumPy array for efficient computation
-    neg_pairs = np.fromiter(
-        itertools.chain.from_iterable(neg_pairs.values()),
-        dtype=np.dtype((np.uint32, 2)),
-        count=neg_total,
-    )
 
     # Compute similarities for positive pairs
     logger.info("Computing positive similarities...")