I have the strong intuition that duckdb can speed up matching by quite a bit (and also the code would be considerably simpler). So I decided to test it out.

EDIT: Current benchmarking code:
```python
from timeit import Timer

setup = """
import pandas as pd

from copairs.matching import Matcher, find_pairs

def simulate_plates(n_compounds, n_replicates, plate_size):
    total = n_compounds * n_replicates

    compounds = []
    plates = []
    wells = []
    for i in range(total):
        compound_id = i % n_compounds
        well_id = i % plate_size
        plate_id = i // plate_size
        compounds.append(f"c{{compound_id}}")
        plates.append(f"p{{plate_id}}")
        wells.append(f"w{{well_id}}")

    dframe = pd.DataFrame({{"c": compounds, "p": plates, "w": wells}})
    return dframe

sameby = ["c", "w"]
diffby = []
SEED = 42
n_replicates = 20
plate_size = 384

dframe = simulate_plates({}, n_replicates, plate_size)
# Load matcher
matcher = Matcher(dframe, dframe.columns, seed=SEED)

"""

n_times = 3
n_repeats = 5
for n_compounds in (10**i for i in range(3, 6)):
    for fn in (
        "find_pairs(dframe,sameby,diffby)",
        "matcher.get_all_pairs(sameby,diffby)",
    ):
        times = Timer(fn, setup=setup.format(n_compounds)).repeat(n_times, n_repeats)
        avg = sum(times) / n_repeats
        print(f"{fn.split('(')[0]},{n_compounds=} has an avg time of {avg:0.3f} secs.")
```

```
matcher.get_all_pairs,n_compounds=1000 has an avg time of 1.760 secs.
matcher.get_all_pairs,n_compounds=10000 has an avg time of 17.325 secs.
matcher.get_all_pairs,n_compounds=100000 has an avg time of 177.668 secs.
find_pairs,n_compounds=1000 has an avg time of 0.061 secs.
find_pairs,n_compounds=10000 has an avg time of 0.148 secs.
find_pairs,n_compounds=100000 has an avg time of 0.561 secs.
```