@afermg 2025-02-14
From a conversation with John, the original Matcher class returns a dictionary so (same1, same2,...) -> (index1,index2). A simple join in duckdb returns a dataframe like this, which may be useful to be passed downstream to the step that calculates distances:


```
┌─────────┬───────┬───────┐
│    c    │ index │ index │
│ varchar │ int64 │ int64 │
├─────────┼───────┼───────┤
│ c0      │     0 │ 28500 │
│ c1      │     1 │ 28501 │
│ c2      │     2 │ 28502 │
│ c3      │     3 │ 28503 │
│ c4      │     4 │ 28504 │
│ c5      │     5 │ 28505 │
│ c6      │     6 │ 28506 │
│ c7      │     7 │ 28507 │
│ c8      │     8 │ 28508 │
│ c9      │     9 │ 28509 │
│ ·       │     · │   ·   │
│ ·       │     · │   ·   │
│ ·       │     · │   ·   │
│ c298    │  1798 │ 22798 │
│ c299    │  1799 │ 22799 │
│ c300    │  1800 │ 22800 │
│ c301    │  1801 │ 22801 │
│ c302    │  1802 │ 22802 │
│ c303    │  1803 │ 22803 │
│ c304    │  1804 │ 22804 │
│ c305    │  1805 │ 22805 │
│ c306    │  1806 │ 22806 │
│ c307    │  1807 │ 22807 │
├─────────┴───────┴───────┤
│ ? rows        3 columns │
└─────────────────────────┘
```