@afermg 2025-02-14
I still need to return a value that can be actually used in downstream analyses.