@afermg 2025-02-14
Based on the code [here](https://github.com/afermg/copairs/blob/main/src/copairs/map/average_precision.py#L192-L197), it seems like it is fine leaving it as an array. I can replace this with to_numpy and that should work just fine.