@afermg 2025-02-14
Actually, that entire section converts back from the nested dictionaries into numpy. We already have it as a table so we are good to go, I may just need to perform the duckdb-> numpy conversion there.