@afermg 2025-02-14
Some notes for self:

There is a bug for the empty case, because the Join is still comparing index 0 vs itself, if we perform same_by we are getting also the duplicate index 0 vs index 0. We may be also still be getting index X = index Y AND index Y = index X, which could be leading to twice the calculations. I set up a script to time both

```python
from timeit import Timer

setup = """
import pandas as pd

from copairs.matching import Matcher, find_pairs

def simulate_plates(n_compounds, n_replicates, plate_size):
    total = n_compounds * n_replicates

    compounds = []
    plates = []
    wells = []
    for i in range(total):
        compound_id = i % n_compounds
        well_id = i % plate_size
        plate_id = i // plate_size
        compounds.append(f"c{compound_id}")
        plates.append(f"p{plate_id}")
        wells.append(f"w{well_id}")

    dframe = pd.DataFrame({"c": compounds, "p": plates, "w": wells})
    return dframe

sameby = ["c", "w"]
diffby = []
SEED = 42
n_replicates = 20
plate_size = 384

dframe = simulate_plates({}, n_replicates, plate_size)
# Load matcher
matcher = Matcher(dframe, dframe.columns, seed=SEED)

"""

n_times = 3
n_repeats = 5
for n_compounds in (10**i for i in range(3, 6)):
    for fn in (
        "find_pairs(dframe,sameby,diffby)",
        "matcher.get_all_pairs(sameby,diffby)",
    ):
        times = Timer(fn, setup=setup.format(n_compounds)).repeat(n_times, n_repeats)
        avg = sum(times) / n_repeats
        print(f"{fn.split('(')[0]},{n_compounds=} has an avg time of {avg:0.3f} secs.")
```