@afermg 2025-02-14
I mean, the numbers above seem to indicate an speedup between  29 and 210x, and it would be larger if we had to deal with a larger dataset. Not to bad of an improvement if I may say so. 

A summary of what my coded added is as follows:
- Added duckdb as a core dependency
- Added ipdb + jupyter as development dependencies (feel free to remove)
- Added a nix flake and lock to make the environment reproducible (feel free to remove)
- Added the copairs.matching.find_pairs function that greatly increases the matching speed.\

I will submit for review now.