@johnarevalo 2025-02-15
Thanks for the PR, it is very exciting to see these improvements! 

One missing part is that the test suite is not really testing the new duckdb impl. because it is still using the `Matcher` impl.

Could you please modify the `Matcher` class, so that the `test_matching*.py` can validate the generated pairs with the new approach? I'm more than happy to get rid of that class, but before that, let's run the tests cases as is. i.e. do not modify the test suite during this transition.

Once we are confident on the tests, we can drop the `Matcher` and update the test suite accordingly.