@afermg 2025-02-15
Most tests adjusted and the type-checking needed was added. 
Tests pass:
```
~/projects/copairs duckdb* amunoz@gpa85-cad
❯ pytest
============================= test session starts ==============================platform linux -- Python 3.11.11, pytest-8.3.4, pluggy-1.5.0
rootdir: /home/amunoz/projects/copairs
configfile: pyproject.toml
plugins: anyio-4.8.0
collected 38 items                                                             

tests/test_build_rank_multilabel.py .                                    [  2%]
tests/test_compute.py ....                                               [ 13%]
tests/test_map.py .......                                                [ 31%]
tests/test_map_filter.py ....                                            [ 42%]
tests/test_matching.py .........                                         [ 65%]
tests/test_matching_any.py ....                                          [ 76%]
tests/test_matching_multilabel.py .....                                  [ 89%]
tests/test_reference_index.py .                                          [ 92%]
tests/test_replicating.py ...                                            [100%]

============================= 38 passed in 41.76s ==============================
```
The only test left is `run_stress_sample_null`. I need to re-read Matcher to see what it is doing, from what I gather it is related to sampling pairs randomly(?). Seems easy to reimplement.